working on sonar

+++ tests/faker_constant.py
@@ -10,7 +10,9 @@ class Provider(BaseProvider):
     def random_int(self, min: int = 0, max: int = 9999, step: int = 1) -> int:
         return 1
 
-    def random_float(self, digits: Optional[int] = None, fix_len: bool = False) -> int:
+    def random_float(
+        self, digits: Optional[int] = None, fix_len: bool = False
+    ) -> float:
         return 1.2
 
     def date(self):
